Broke the assertions file into sub-parts
