+++ kakao/src/main/kotlin/com/agoda/kakao/assertions/AdapterAssertions.kt
@@ -1,0 +1,12 @@
+package com.agoda.kakao.assertions
+
+import android.support.test.espresso.ViewInteraction
+
+/**
+ * Provides assertions for view with adapters
+ *
+ * @see RecyclerAdapterAssertions
+ */
+interface AdapterAssertions {
+    val view: ViewInteraction
+}
