+++ kakao/src/main/kotlin/com/agoda/kakao/assertions/CheckableAssertions.kt
@@ -1,0 +1,25 @@
+package com.agoda.kakao.assertions
+
+import android.support.test.espresso.assertion.ViewAssertions
+import android.support.test.espresso.matcher.ViewMatchers
+
+/**
+ * Provides checkable based assertions for views
+ */
+interface CheckableAssertions : BaseAssertions {
+    /**
+     * Checks if the view is checked
+     */
+    fun isChecked() {
+        view.check(ViewAssertions.matches(
+                ViewMatchers.isChecked()))
+    }
+
+    /**
+     * Checks if the view is not checked
+     */
+    fun isNotChecked() {
+        view.check(ViewAssertions.matches(
+                ViewMatchers.isNotChecked()))
+    }
+}
