+++ kakao/src/main/kotlin/com/agoda/kakao/assertions/ProgressBarAssertions.kt
@@ -1,0 +1,18 @@
+package com.agoda.kakao.assertions
+
+import android.support.test.espresso.assertion.ViewAssertions
+import com.agoda.kakao.matchers.ProgressMatcher
+
+/**
+ * Provides assertions for progress bar
+ */
+interface ProgressBarAssertions : BaseAssertions {
+    /**
+     *  Checks if view has number of progress as expected
+     *
+     *  @param number progress as expected
+     */
+    fun hasProgress(number: Int) {
+        view.check(ViewAssertions.matches(ProgressMatcher(number)))
+    }
+}
