+++ kakao/src/main/kotlin/com/agoda/kakao/assertions/SwipeRefreshLayoutAssertions.kt
@@ -1,0 +1,23 @@
+package com.agoda.kakao.assertions
+
+import android.support.test.espresso.assertion.ViewAssertions
+import com.agoda.kakao.matchers.SwipeRefreshLayoutMatcher
+
+/**
+ * Provides assertion for SwipeRefreshLayout
+ */
+interface SwipeRefreshLayoutAssertions : BaseAssertions {
+    /**
+     * Checks if the SwipeRefreshLayout is refreshing
+     */
+    fun isRefreshing() {
+        view.check(ViewAssertions.matches(SwipeRefreshLayoutMatcher(true)))
+    }
+
+    /**
+     * Checks if the SwipeRefreshLayout is not refreshing
+     */
+    fun isNotRefreshing() {
+        view.check(ViewAssertions.matches(SwipeRefreshLayoutMatcher(false)))
+    }
+}
